refactor(renderer): App 事件订阅抽 use-session-event-bridge，bootstrap 独立精简 effect
